Rename category to segment in classes related to SiteCreation
